+++ contracts/crowdsale/CHANGELOG.md
@@ -6,9 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2022-11-28
+### Changed
+- Updated `gstd`, `gtest`, `gear-wasm-builder` to the `stable` branch.
+- Renamed `crowdsale-ico` to `crowdsale`.
+
 ## [0.1.0] - 2022-10-14
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/gear-dapps/crowdsale-ico/compare/0.1.0...HEAD
-[0.1.0]: https://github.com/gear-dapps/crowdsale-ico/compare/ee7cf43...0.1.0
+[Unreleased]: https://github.com/gear-dapps/crowdsale/compare/0.1.1...HEAD
+[0.1.1]: https://github.com/gear-dapps/crowdsale/compare/0.1.0...0.1.1
+[0.1.0]: https://github.com/gear-dapps/crowdsale/compare/ee7cf43...0.1.0